// src/serverApi/v3/models/card-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TextElementResponse } from './text-element-response';
import { TimestampsResponse } from './timestamps-response';
import { VisibilitySettingsResponse } from './visibility-settings-response';

/**
 * 
 * @export
 * @interface CardResponse
 */
export interface CardResponse {
    /**
     * 
     * @type {string}
     * @memberof CardResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof CardResponse
     */
    title?: string;
    /**
     * 
     * @type {number}
     * @memberof CardResponse
     */
    height: number;
    /**
     * 
     * @type {Array<TextElementResponse>}
     * @memberof CardResponse
     */
    elements: Array<TextElementResponse>;
    /**
     * 
     * @type {VisibilitySettingsResponse}
     * @memberof CardResponse
     */
    visibilitySettings: VisibilitySettingsResponse;
    /**
     * 
     * @type {TimestampsResponse}
     * @memberof CardResponse
     */
    timestamps: TimestampsResponse;
}